Move low-level hashing to crypto.py
